Remove pie chart and begin consolidation of bar/line charts
