--- src/components/11-infographics/pie.njk
@@ -1,20 +1,0 @@
-<div class="usa-chart usa-chart-pie applepie-chart">
-  <table class="usa-chart-data" data-total="100%">
-    <tr class="usa-chart-item">
-      <td class="usa-chart-label">Thing one</td>
-      <td class="usa-chart-value">8%</td>
-    </tr>
-    <tr class="usa-chart-item">
-      <td class="usa-chart-label">Thing two</td>
-      <td class="usa-chart-value">12%</td>
-    </tr>
-    <tr class="usa-chart-item">
-      <td class="usa-chart-label">Thing three</td>
-      <td class="usa-chart-value">26%</td>
-    </tr>
-    <tr class="usa-chart-item">
-      <td class="usa-chart-label">Thing four</td>
-      <td class="usa-chart-value">54%</td>
-    </tr>
-  </table>
-</div>
