+++ src/stylesheets/components/_chart.scss
@@ -1,0 +1,49 @@
+$speed: .25s;
+$infographic-colors:
+  $color-blue
+  $color-aqua
+  $color-green
+  $color-purple
+  $color-cool-blue
+  $color-gold
+  $color-red;
+
+// Chart styles
+
+.usa-chart-labelText,
+.usa-chart-labelValue {
+  font-size: calc(3px - .1vw);
+  text-anchor: middle;
+}
+
+.usa-chart-labelValue-invert {
+  fill: #fff;
+}
+
+.usa-chart-gridLabel {
+  font-size: calc(3px - .1vw);
+  fill: #999;
+  text-anchor: end;
+}
+
+.usa-chart-gridLine {
+  stroke: #eee;
+  stroke-width: .2;
+}
+
+.usa-chart-bars,
+.usa-chart-point {
+  fill: nth($infographic-colors, 1);
+  transition: all $speed ease;
+}
+.usa-chart-path {
+  stroke-width: .5px;
+  stroke: nth($infographic-colors, 1);
+  fill: none;
+}
+
+.usa-chart-axis {
+  stroke: #aaa;
+  stroke-width: .2;
+  fill: none;
+}
