--- src/stylesheets/components/_infographics.scss
@@ -1,134 +1,0 @@
-$speed: .25s;
-$infographic-colors:
-  $color-blue
-  $color-aqua
-  $color-green
-  $color-purple
-  $color-cool-blue
-  $color-gold
-  $color-red;
-
-// Bar and line charts specific styles
-
-.usa-chart-labelText,
-.usa-chart-labelValue {
-  font-size: calc(3px - .1vw);
-  text-anchor: middle;
-}
-
-.usa-chart-labelValue-invert {
-  fill: #fff;
-}
-
-.usa-chart-gridLabel {
-  font-size: calc(3px - .1vw);
-  fill: #999;
-  text-anchor: end;
-}
-
-.usa-chart-gridLine {
-  stroke: #eee;
-  stroke-width: .2;
-}
-
-.usa-chart-bars,
-.usa-chart-point {
-  fill: nth($infographic-colors, 1);
-  transition: all $speed ease;
-}
-.usa-chart-path {
-  stroke-width: .5px;
-  stroke: nth($infographic-colors, 1);
-  fill: none;
-}
-
-.usa-chart-axis {
-  stroke: #aaa;
-  stroke-width: .2;
-  fill: none;
-}
-
-// Pie chart specific styles
-.usa-chart-pie.rendered {
-  display: flex;
-  align-items: center;
-
-  td {
-    border: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  .usa-chart-item {
-    border-bottom: 2px solid transparent;
-  }
-
-  .usa-chart-label {
-    padding-right: .5em;
-    position: relative;
-    transition: border-bottom $speed ease;
-  }
-
-  .usa-chart-data,
-  .usa-chart-svg {
-    flex: 1;
-    margin: 0;
-  }
-
-  .usa-chart-svg {
-    flex: 2;
-    max-width: 66.7%;
-  }
-}
-
-.usa-chart-value {
-  font-weight: 700;
-}
-
-.usa-chart-arc {
-  stroke: currentColor;
-  stroke-width: 8%;
-  animation: $speed 'usa-chart-pie' 1 ease;
-  fill: none;
-  transition: all $speed ease;
-
-  &:hover,
-  &.hover {
-    stroke-width: 12%;
-  }
-}
-
-// Loop through and write graph colors
-@for $i from 1 through length($infographic-colors) {
-
-  .usa-chart-arc:nth-child(#{length($infographic-colors)}n+#{$i}),
-  .usa-chart-item:nth-child(#{length($infographic-colors)}n+#{$i}) .usa-chart-value {
-    color: nth($infographic-colors, $i);
-  }
-
-  .usa-chart-item:nth-child(#{length($infographic-colors)}n+#{$i}):hover,
-  .usa-chart-item:nth-child(#{length($infographic-colors)}n+#{$i}).hover {
-    border-bottom-color: nth($infographic-colors, $i);
-  }
-
-  // Ensure last color doesn't clash with the first
-  @if ($i == 1) {
-
-    .usa-chart-arc:nth-child(#{length($infographic-colors)}n+#{$i}):last-child,
-    .usa-chart-item:nth-child(#{length($infographic-colors)}n+#{$i}):last-child .usa-chart-value {
-      color: nth($infographic-colors, round(length($infographic-colors)/2));
-    }
-
-    .usa-chart-item:nth-child(#{length($infographic-colors)}n+#{$i}):last-child:hover,
-    .usa-chart-item:nth-child(#{length($infographic-colors)}n+#{$i}):last-child.hover {
-      border-bottom-color: nth($infographic-colors, round(length($infographic-colors)/2));
-    }
-  }
-}
-
-@keyframes 'usa-chart-pie' {
-  0% {
-    stroke-dasharray: 0, (100% * 3.1415926535897932384626433832795/2);
-    stroke-dashoffset: 0;
-  }
-}
